# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Simple server in order to create and mantain CGA Survey"
  version: "1.0.0"
  title: "CGA Survey"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "a@a.com"
  license:
    name: "MIT"
    url: "https://github.com/TrentoCrowdAI/CGA/blob/master/LICENSE"
host: "cga-mock-server.herokuapp.com"
basePath: "/v1"
...
schemes:
- "https"
paths:
...
  /surveys:
    get:
      tags:
      - "survey"
      summary: "Get all survey"
      description: "Performed by a user in order to get all surveys"
      operationId: "getSurveys"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/survey"
        400:
          description: "Bad request"
      x-swagger-router-controller: "Survey"
    post:
      tags:
      - "survey"
      summary: "Add a new survey"
      description: "Performed by a user when he wants to create a new survey in order\
        \ to collect patient data"
      operationId: "addSurvey"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "survey"
        description: "Survey object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/survey_1"
      responses:
        201:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/survey"
        400:
          description: "Bad Request"
      x-swagger-router-controller: "Survey"
  /surveys/{id}:
    get:
      tags:
      - "survey"
      summary: "Get an survey by the id"
      description: "Performed by users in order to get specific survey"
      operationId: "getSurvey"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The survey id"
        required: true
        type: "integer"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/survey"
        400:
          description: "Invalid id"
        404:
          description: "Survey not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "Survey"
    put:
      tags:
      - "survey"
      summary: "Update an existing survey"
      description: "Performed by a user in order to update the survey"
      operationId: "updateSurvey"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The survey id"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "Survey object that needs to be updated"
        required: true
        schema:
          $ref: "#/definitions/body_2"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/survey"
        400:
          description: "Invalid id"
        404:
          description: "Survey not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "Survey"
    delete:
      tags:
      - "survey"
      summary: "Delete a survey by id"
      description: "Performed by the user in order to delete the survey from the system"
      operationId: "deleteSurvey"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The survey id"
        required: true
        type: "integer"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/survey"
        400:
          description: "Invalid id"
        404:
          description: "Survey not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "Survey"
  /surveyComponents:
    get:
      tags:
      - "surveyComponent"
      summary: "Get all surveyComponent"
      description: "Performed by a user in order to get all surveyComponent"
      operationId: "getAllSurveyComponents"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/surveyComponent"
        400:
          description: "Bad request"
      x-swagger-router-controller: "SurveyComponent"
    post:
      tags:
      - "surveyComponent"
      summary: "Add a new SurveyComponent"
      description: "Performed by a user when he wants to create a new surveyComponent\
        \ in order to collect patient data"
      operationId: "addSurveyComponent"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "SurveyComponent"
        description: "SurveyComponent object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/SurveyComponent"
      responses:
        201:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/surveyComponent"
        400:
          description: "Bad Request"
      x-swagger-router-controller: "SurveyComponent"
  /surveyComponents/{id}:
    get:
      tags:
      - "surveyComponent"
      summary: "Get a SurveyComponent by the id"
      description: "Performed by users in order to get specific SurveyComponent"
      operationId: "getSurveyComponent"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The SurveyComponent id"
        required: true
        type: "integer"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/surveyComponent"
        400:
          description: "Invalid id"
        404:
          description: "SurveyComponent not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "SurveyComponent"
    put:
      tags:
      - "surveyComponent"
      summary: "Update an existing SurveyComponent"
      description: "Performed by a user in order to update the surveyComponent"
      operationId: "updateSurveyComponent"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The SurveyComponent id"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "SurveyComponent object that needs to be updated"
        required: true
        schema:
          $ref: "#/definitions/body_3"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/surveyComponent"
        400:
          description: "Invalid id"
        404:
          description: "SurveyComponent not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "SurveyComponent"
    delete:
      tags:
      - "surveyComponent"
      summary: "Delete a SurveyItem by id"
      description: "Performed by the user in order to delete the SurveyItem from the\
        \ system"
      operationId: "deleteSurveyComponent"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The SurveyComponent id"
        required: true
        type: "integer"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/surveyComponent"
        400:
          description: "Invalid id"
        404:
          description: "SurveyComponent not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "SurveyComponent"
...
  /rules:
    get:
      tags:
      - "rule"
      summary: "Get all Rules"
      description: "Performed by a user in order to get all Rules"
      operationId: "getAllRules"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/rule"
        400:
          description: "Bad request"
      x-swagger-router-controller: "Rule"
    post:
      tags:
      - "rule"
      summary: "Add a new Rule"
      description: "Performed by a user when he wants to create a new Rule"
      operationId: "addRule"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "rule"
        description: "Rule object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/rule_1"
      responses:
        201:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/rule"
        400:
          description: "Bad Request"
      x-swagger-router-controller: "Rule"
  /rules/{id}:
    get:
      tags:
      - "rule"
      summary: "Get a Rule by the id"
      description: "Performed by users in order to get specific Rule"
      operationId: "getRule"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The Rule id"
        required: true
        type: "integer"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/rule"
        400:
          description: "Invalid id"
        404:
          description: "Rule not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "Rule"
    put:
      tags:
      - "rule"
      summary: "Update an existing Rule"
      description: "Performed by a user in order to update the Rule"
      operationId: "updateRule"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The Rule id"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "Rule object that needs to be updated"
        required: true
        schema:
          $ref: "#/definitions/body_7"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/rule"
        400:
          description: "Invalid id"
        404:
          description: "Rule not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "Rule"
    delete:
      tags:
      - "rule"
      summary: "Delete a Rule by id"
      description: "Performed by the user in order to delete the Rule from the system"
      operationId: "deleteRule"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The Rule id"
        required: true
        type: "integer"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/rule"
        400:
          description: "Invalid id"
        404:
          description: "Rule not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "Rule"
...
  /subjects:
    get:
      tags:
      - "subject"
      summary: "Get all subject"
      description: "Performed by a user in order to get all Subject"
      operationId: "getAllSubjects"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/subject"
        400:
          description: "Bad request"
      x-swagger-router-controller: "Subject"
    post:
      tags:
      - "subject"
      summary: "Add a new subject"
      description: "Operation to call in order to create a new Subject in the system"
      operationId: "addSubject"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "subject"
        description: "Subject object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/subject_1"
      responses:
        201:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/subject"
        400:
          description: "Bad Request"
      x-swagger-router-controller: "Subject"
  /subject/{id}:
    get:
      tags:
      - "subject"
      summary: "Get a Subject by the id"
      description: "Performed by users in order to get specific Subject"
      operationId: "getSubject"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The Subject id"
        required: true
        type: "integer"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/subject"
        400:
          description: "Invalid id"
        404:
          description: "Subject not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "Subject"
    put:
      tags:
      - "subject"
      summary: "Update an existing Subject"
      description: "Performed by a user in order to update the Subject"
      operationId: "updateSubject"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The Subject id"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "Subject object that needs to be updated"
        required: true
        schema:
          $ref: "#/definitions/body_11"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/subject"
        400:
          description: "Invalid id"
        404:
          description: "Subject not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "Subject"
    delete:
      tags:
      - "subject"
      summary: "Delete a Subject by id"
      description: "Performed by the user in order to delete the Subject from the\
        \ system"
      operationId: "deleteSubject"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The Subject id"
        required: true
        type: "integer"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/subject"
        400:
          description: "Invalid id"
        404:
          description: "Subject not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "Subject"
definitions:
...
  survey:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Number that identify the survey in the system"
      name:
        type: "string"
        description: "Name assigned to the survey"
      description:
        type: "string"
        description: "String that describe the survey"
      components:
        type: "array"
        description: "SurveyComponent that compose a survey"
        items:
          $ref: "#/definitions/surveyComponent"
    example:
      components:
      - name: "name"
        id: 5
        items:
        - validators:
          - id: 2
            type: "type"
            value: "value"
          - id: 2
            type: "type"
            value: "value"
          options:
          - id: 7
            type: "type"
            value: "value"
            labels:
            - values: "values"
              language: "language"
              id: 2
            - values: "values"
              language: "language"
              id: 2
          - id: 7
            type: "type"
            value: "value"
            labels:
            - values: "values"
              language: "language"
              id: 2
            - values: "values"
              language: "language"
              id: 2
          description: "description"
          rules:
          - item: 3
            condition: "condition"
            id: 9
          - item: 3
            condition: "condition"
            id: 9
          id: 5
          type: "type"
          flowConstraints:
          - typology: "typology"
            condition: "condition"
            id: 4
            value: "value"
          - typology: "typology"
            condition: "condition"
            id: 4
            value: "value"
          labels:
          - values: "values"
            language: "language"
            id: 2
          - values: "values"
            language: "language"
            id: 2
        - validators:
          - id: 2
            type: "type"
            value: "value"
          - id: 2
            type: "type"
            value: "value"
          options:
          - id: 7
            type: "type"
            value: "value"
            labels:
            - values: "values"
              language: "language"
              id: 2
            - values: "values"
              language: "language"
              id: 2
          - id: 7
            type: "type"
            value: "value"
            labels:
            - values: "values"
              language: "language"
              id: 2
            - values: "values"
              language: "language"
              id: 2
          description: "description"
          rules:
          - item: 3
            condition: "condition"
            id: 9
          - item: 3
            condition: "condition"
            id: 9
          id: 5
          type: "type"
          flowConstraints:
          - typology: "typology"
            condition: "condition"
            id: 4
            value: "value"
          - typology: "typology"
            condition: "condition"
            id: 4
            value: "value"
          labels:
          - values: "values"
            language: "language"
            id: 2
          - values: "values"
            language: "language"
            id: 2
      - name: "name"
        id: 5
        items:
        - validators:
          - id: 2
            type: "type"
            value: "value"
          - id: 2
            type: "type"
            value: "value"
          options:
          - id: 7
            type: "type"
            value: "value"
            labels:
            - values: "values"
              language: "language"
              id: 2
            - values: "values"
              language: "language"
              id: 2
          - id: 7
            type: "type"
            value: "value"
            labels:
            - values: "values"
              language: "language"
              id: 2
            - values: "values"
              language: "language"
              id: 2
          description: "description"
          rules:
          - item: 3
            condition: "condition"
            id: 9
          - item: 3
            condition: "condition"
            id: 9
          id: 5
          type: "type"
          flowConstraints:
          - typology: "typology"
            condition: "condition"
            id: 4
            value: "value"
          - typology: "typology"
            condition: "condition"
            id: 4
            value: "value"
          labels:
          - values: "values"
            language: "language"
            id: 2
          - values: "values"
            language: "language"
            id: 2
        - validators:
          - id: 2
            type: "type"
            value: "value"
          - id: 2
            type: "type"
            value: "value"
          options:
          - id: 7
            type: "type"
            value: "value"
            labels:
            - values: "values"
              language: "language"
              id: 2
            - values: "values"
              language: "language"
              id: 2
          - id: 7
            type: "type"
            value: "value"
            labels:
            - values: "values"
              language: "language"
              id: 2
            - values: "values"
              language: "language"
              id: 2
          description: "description"
          rules:
          - item: 3
            condition: "condition"
            id: 9
          - item: 3
            condition: "condition"
            id: 9
          id: 5
          type: "type"
          flowConstraints:
          - typology: "typology"
            condition: "condition"
            id: 4
            value: "value"
          - typology: "typology"
            condition: "condition"
            id: 4
            value: "value"
          labels:
          - values: "values"
            language: "language"
            id: 2
          - values: "values"
            language: "language"
            id: 2
      name: "name"
      description: "description"
      id: 1
  surveyComponent:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Number that identify the SurveyComponent"
      name:
        type: "string"
        description: "Name thah describe the SurveyComponent"
      items:
        type: "array"
        description: "Items that compose a surveyComponent"
        items:
          $ref: "#/definitions/surveyItem"
    example:
      name: "name"
      id: 5
      items:
      - validators:
        - id: 2
          type: "type"
          value: "value"
        - id: 2
          type: "type"
          value: "value"
        options:
        - id: 7
          type: "type"
          value: "value"
          labels:
          - values: "values"
            language: "language"
            id: 2
          - values: "values"
            language: "language"
            id: 2
        - id: 7
          type: "type"
          value: "value"
          labels:
          - values: "values"
            language: "language"
            id: 2
          - values: "values"
            language: "language"
            id: 2
        description: "description"
        rules:
        - item: 3
          condition: "condition"
          id: 9
        - item: 3
          condition: "condition"
          id: 9
        id: 5
        type: "type"
        flowConstraints:
        - typology: "typology"
          condition: "condition"
          id: 4
          value: "value"
        - typology: "typology"
          condition: "condition"
          id: 4
          value: "value"
        labels:
        - values: "values"
          language: "language"
          id: 2
        - values: "values"
          language: "language"
          id: 2
      - validators:
        - id: 2
          type: "type"
          value: "value"
        - id: 2
          type: "type"
          value: "value"
        options:
        - id: 7
          type: "type"
          value: "value"
          labels:
          - values: "values"
            language: "language"
            id: 2
          - values: "values"
            language: "language"
            id: 2
        - id: 7
          type: "type"
          value: "value"
          labels:
          - values: "values"
            language: "language"
            id: 2
          - values: "values"
            language: "language"
            id: 2
        description: "description"
        rules:
        - item: 3
          condition: "condition"
          id: 9
        - item: 3
          condition: "condition"
          id: 9
        id: 5
        type: "type"
        flowConstraints:
        - typology: "typology"
          condition: "condition"
          id: 4
          value: "value"
        - typology: "typology"
          condition: "condition"
          id: 4
          value: "value"
        labels:
        - values: "values"
          language: "language"
          id: 2
        - values: "values"
          language: "language"
          id: 2
...
  rule:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Number that identify the label"
      item:
        type: "integer"
        description: "id of the previous item in order to check its value"
      condition:
        type: "string"
        description: "Condition to verify"
    example:
      item: 3
      condition: "condition"
      id: 9
...
  survey_1:
    type: "object"
    properties:
      name:
        type: "string"
        description: "Name assigned to the survey"
      description:
        type: "string"
        description: "String that describe the survey"
      components:
        type: "array"
        description: "SurveyComponent that compose a survey"
        items:
          $ref: "#/definitions/surveyComponent"
  body_2:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Number that identify the survey in the system"
      name:
        type: "string"
        description: "Name assigned to the survey"
      description:
        type: "string"
        description: "String that describe the survey"
      components:
        type: "array"
        description: "SurveyComponent that compose a survey"
        items:
          $ref: "#/definitions/surveyComponent"
  SurveyComponent:
    type: "object"
    properties:
      name:
        type: "string"
        description: "Name thah describe the SurveyComponent"
      items:
        type: "array"
        description: "Items that compose a surveyComponent"
        items:
          $ref: "#/definitions/surveyItem"
    example:
      name: "name"
      items:
      - validators:
        - id: 2
          type: "type"
          value: "value"
        - id: 2
          type: "type"
          value: "value"
        options:
        - id: 7
          type: "type"
          value: "value"
          labels:
          - values: "values"
            language: "language"
            id: 2
          - values: "values"
            language: "language"
            id: 2
        - id: 7
          type: "type"
          value: "value"
          labels:
          - values: "values"
            language: "language"
            id: 2
          - values: "values"
            language: "language"
            id: 2
        description: "description"
        rules:
        - item: 3
          condition: "condition"
          id: 9
        - item: 3
          condition: "condition"
          id: 9
        id: 5
        type: "type"
        flowConstraints:
        - typology: "typology"
          condition: "condition"
          id: 4
          value: "value"
        - typology: "typology"
          condition: "condition"
          id: 4
          value: "value"
        labels:
        - values: "values"
          language: "language"
          id: 2
        - values: "values"
          language: "language"
          id: 2
      - validators:
        - id: 2
          type: "type"
          value: "value"
        - id: 2
          type: "type"
          value: "value"
        options:
        - id: 7
          type: "type"
          value: "value"
          labels:
          - values: "values"
            language: "language"
            id: 2
          - values: "values"
            language: "language"
            id: 2
        - id: 7
          type: "type"
          value: "value"
          labels:
          - values: "values"
            language: "language"
            id: 2
          - values: "values"
            language: "language"
            id: 2
        description: "description"
        rules:
        - item: 3
          condition: "condition"
          id: 9
        - item: 3
          condition: "condition"
          id: 9
        id: 5
        type: "type"
        flowConstraints:
        - typology: "typology"
          condition: "condition"
          id: 4
          value: "value"
        - typology: "typology"
          condition: "condition"
          id: 4
          value: "value"
        labels:
        - values: "values"
          language: "language"
          id: 2
        - values: "values"
          language: "language"
          id: 2
  body_3:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Number that identify the SurveyComponent"
      name:
        type: "string"
        description: "Name thah describe the SurveyComponent"
      items:
        type: "array"
        description: "Items that compose a surveyComponent"
        items:
          $ref: "#/definitions/surveyItem"
...
  rule_1:
    type: "object"
    properties:
      item:
        type: "integer"
        description: "id of the previous item in order to checkoì its value"
      condition:
        type: "string"
        description: "Condition to verify"
...